themes: Fix incorrect GradientActiveTitle colour in "Warm" themes
The GradientActiveTitle colour in "Warm" version of the Windows Standard
theme retained the default value from the original theme instead of
having its R and B colours swapped. Thus, swap the colours, so that they
actually combine well with the ActiveTitle colour.

On a side note, this change has no effect when used with high contrast
themes, as those apply a uniform ActiveTitle colour to their windows.
However, the GradientActiveTitle colour is still used with actual
classic themes, either in older Windows versions, or when the classic
theme has been enabled in Windows 8+ with hacks.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>